+++ packages/ubdev/public/configurator/ConfiguratorComponent.vue
@@ -1,14 +1,31 @@
 <template>
-  <el-tabs type="border-card" v-if="metaObject && schemaObject" @tab-click="reloadCodeMirrorData">
+  <el-tabs
+    v-if="metaObject && schemaObject"
+    type="border-card"
+    @tab-click="reloadCodeMirrorData"
+  >
     <el-tab-pane label="Entity">
-      <objCardComponent :metaObject="metaObject" :fileName="fileName" :schema="schemeAttributes"></objCardComponent>
+      <objCardComponent
+        :meta-object="metaObject"
+        :file-name="fileName"
+        :schema="schemeAttributes"
+      />
     </el-tab-pane>
     <el-tab-pane label="Mixins">
-      <mixinsCardComponent :mixins="metaObject.mixins"
-                           :schema="schemaObject.properties.mixins.properties"></mixinsCardComponent>
+      <mixinsCardComponent
+        :mixins="metaObject.mixins"
+        :schema="schemaObject.properties.mixins.properties"
+      />
     </el-tab-pane>
-    <el-tab-pane id="sourceTab" label="Source">
-      <el-input :id="_uid" type="textarea" v-model="outputJson"></el-input>
+    <el-tab-pane
+      id="sourceTab"
+      label="Source"
+    >
+      <el-input
+        :id="_uid"
+        v-model="outputJson"
+        type="textarea"
+      />
     </el-tab-pane>
   </el-tabs>
 </template>
@@ -16,6 +33,7 @@
 <script>
 const objCardComponent = require('./ObjectCardComponent.vue')
 const mixinsCardComponent = require('./MixinsCardComponent.vue')
+const {UBDomain} = require('@unitybase/cs-shared')
 
 module.exports = {
   props: {
@@ -37,19 +55,20 @@ module.exports = {
     },
     outputJson: {
       get: function () {
-        if (this.metaObject) {
-          let minObj = {...this.metaObject}
-          minObj.attributes.forEach((attribute) => {
-            Object.keys(attribute).forEach((key) => {
-              let schemeAttr = this.schemeAttributes[key]
-              if (!schemeAttr) this.$delete(attribute, key)
-              if (schemeAttr && schemeAttr.default && attribute[key] === schemeAttr.default) {
-                this.$delete(attribute, key)
-              }
-            }, this)
-          }, this)
-          return JSON.stringify(minObj, this.jsonReplacer, 2)
-        }
+        if (!this.metaObject) return
+        let e = new UBDomain.UBEntity(this.metaObject)
+        return JSON.stringify(e.asJSON(), null, '  ')
+        // let minObj = { ...this.metaObject }
+        // minObj.attributes.forEach((attribute) => {
+        //   Object.keys(attribute).forEach((key) => {
+        //     let schemeAttr = this.schemeAttributes[key]
+        //     if (!schemeAttr) this.$delete(attribute, key)
+        //     if (schemeAttr && schemeAttr.default && attribute[key] === schemeAttr.default) {
+        //       this.$delete(attribute, key)
+        //     }
+        //   }, this)
+        // }, this)
+        // return JSON.stringify(minObj, this.jsonReplacer, 2)
       },
       set: function (newValue) {
         this.initMetaObject(newValue)
@@ -67,33 +86,31 @@ module.exports = {
   },
   methods: {
     initCodeMirror () {
-      if (!this.codeMirror) {
-        System.import('@unitybase/codemirror-full').then((CodeMirror) => {
-          var el = document.getElementById(this._uid)
-          if (el) {
-            this.codeMirror = CodeMirror.fromTextArea(el, {
-              mode: 'javascript',
-              lineNumbers: true,
-              lint: Object.assign({asi: true, esversion: 6}, $App.connection.appConfig.uiSettings.adminUI.linter),
-              readOnly: false,
-              tabSize: 2,
-              highlightSelectionMatches: {annotateScrollbar: true},
-              matchBrackets: true,
-              autofocus: true,
-              foldGutter: true,
-              gutters: ['CodeMirror-linenumbers', 'CodeMirror-foldgutter', 'CodeMirror-lint-markers'],
-              extraKeys: {
-                'Ctrl-Space': 'autocomplete'
-              }
-            })
-            this.codeMirror.on('change', function (cmInstance) {
-              if (this.outputJson !== cmInstance.getValue()) this.outputJson = cmInstance.getValue()
-              this.refreshCodeMirror()
-            }.bind(this))
-            this.refreshCodeMirror()
+      if (this.codeMirror) return
+      System.import('@unitybase/codemirror-full').then((CodeMirror) => {
+        var el = document.getElementById(this._uid)
+        if (!el) return
+        this.codeMirror = CodeMirror.fromTextArea(el, {
+          mode: 'javascript',
+          lineNumbers: true,
+          lint: Object.assign({ asi: true, esversion: 6 }, $App.connection.appConfig.uiSettings.adminUI.linter),
+          readOnly: false,
+          tabSize: 2,
+          highlightSelectionMatches: { annotateScrollbar: true },
+          matchBrackets: true,
+          autofocus: true,
+          foldGutter: true,
+          gutters: ['CodeMirror-linenumbers', 'CodeMirror-foldgutter', 'CodeMirror-lint-markers'],
+          extraKeys: {
+            'Ctrl-Space': 'autocomplete'
           }
-        }, this)
-      }
+        })
+        this.codeMirror.on('change', function (cmInstance) {
+          if (this.outputJson !== cmInstance.getValue()) this.outputJson = cmInstance.getValue()
+          this.refreshCodeMirror()
+        }.bind(this))
+        this.refreshCodeMirror()
+      })
     },
     jsonReplacer (key, value) {
       if (
@@ -117,15 +134,14 @@ module.exports = {
     initMetaObject (json) {
       let valid = true
       try {
-        this.metaObject = typeof json === 'string' ? JSON.parse(json) : {...json}
+        this.metaObject = typeof json === 'string' ? JSON.parse(json) : { ...json }
       } catch (e) {
         valid = false
       }
       if (valid && this.metaObject) {
         if (!Array.isArray(this.metaObject.attributes)) {
           Object.keys(this.metaObject.attributes).forEach((propertyName) => {
-            if (!this.metaObject.attributes[propertyName]['name'])
-              this.$set(this.metaObject.attributes[propertyName], 'name', propertyName)
+            if (!this.metaObject.attributes[propertyName]['name']) { this.$set(this.metaObject.attributes[propertyName], 'name', propertyName) }
           })
           this.$set(this.metaObject, 'attributes', Object.values(this.metaObject.attributes))
         }